Fixes and update pipeline

+++ include/pipeline/task.hpp
@@ -15,11 +15,10 @@ enum TaskType { FORWARD, BACKWARD };
 template <typename T = float> struct Task {
   TaskType type;
   Tensor<T> data;
-  size_t micro_batch_id;
+  int micro_batch_id;
 
   Task() = default;
-
-  Task(TaskType t, const Tensor<T> &d, size_t mb_id)
+  Task(TaskType t, const Tensor<T> &d, int mb_id)
       : type(t), data(d), micro_batch_id(mb_id) {}
 
   // Copy constructor
